Use views rather than CPU-based hashing for 8- and 16-bit Labels data (#6467)
This is follow-up to #6411.

When labels data are int8, uint8, int16, or uint16, no copy is made —
the data are sent directly as a uint8 or uint16 view to the GPU for
rendering.

Higher bit depth integer data is first converted on the CPU to uint8,
uint16, or float32 (depending on the number of colors in the colormap)
using the formula `texture = (values - 1) % ncolors + 1` — but with the
background value always mapping to 0.

---------

Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>
Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>