Minimal changes for mlx (and jax) compatibility for Image layers (#6553)
These two changes are sufficient to allow napari.imshow to work with
arrays from mlx, Apple's new array library:

1. The function `range_to_decimals` had a tiny hardcoded hack for
tensorstore array dtype. But we have since then implemented the
normalize_dtype function that works with a much broader range of
arrays. In this PR, we change to use normalize_dtype, which works
out of the box with mlx arrays.
2. At one point we use `Image.data.shape` to try to make a set. It turns
out that mlx arrays return a list, rather than a tuple, when checking the
shape, so this PR coerces .shape to a tuple to make sure things are
working.

I don't want to add an mlx test dependency because it's such a young
library, but these changes are very inoffensive so I hope we can just
merge them. You can test that this works locally with:

```python
import napari
import mlx.core as mx


ran = mx.random.uniform(shape=(5, 10, 10))

viewer, layer = napari.imshow(ran)

if __name__ == '__main__':
    napari.run()
```

This works great with these changes! 😊