Added support for features in surface layers (#6515)
# References and relevant issues
Partial fix of #5205 

# Description
This PR adds the most basic form of support for adding the `features` to
the `Surface` layer. In essence, I copied the respective functionality
from the `Points` layer into the code for the Surface layer. *Note*: As
discussed in the [above issue](#5205) this currently modifies the
Surface layer only in so far as it does not throw an error if a
`features` keyword is passed to the instantiation, but it does not hook
up the coloring scheme to the features table.

---------

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>
Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>