Fix label direct mode for installation without numba (#6571)
# Description

When working on fix bug in npe2 I spot that napari tests in npe2
repository are failing because of lack of numba.

In the context of one of the tests, it needs to be skipped when numba is
not installed.

However, the second failure reveals a real bug in codebase.  

This PR also modifies `pip` workflow to be used to test numbaless case.