Fix points `changed` event emission (#6611)
# References and relevant issues
This PR fixes an issue reported on
[forum.image.sc](https://forum.image.sc/t/napari-how-to-distinguish-between-shape-added-removed-by-user-and-back-end/78529/12).
It s one of the latest posts.

# Description
This PR moves the event emission when a point has changed from `_move`
in the `Points` layer to `select` in `_points_mouse_bindings`. This
because in `_move` events where emitted during moving and then the last
event to be emitted when the dragging stops and the user releases the
mouse does not have `selected_indices`. For example below, move the
point and then check console output once drag is finished. You see the
action is `CHANGED` but `data_indices` is empty while it should be `(0)`

```python
import napari

def check_call(event):
    print(event.action)
    print(event.data_indices)


viewer = napari.Viewer()
layer = viewer.add_points([5,5])
layer.events.data.connect(check_call)
napari.run()
```

The event now is only emitted when dragging stops and the output for the
example is now as expected. The downside is that it is more challenging
to test this as `Mock` with `layer.data.events` does not emit anything
now by purely using the `_move` method. This still needs to be fixed.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>