Update conf.py to display the announcement banner (#342)
# References and relevant issues

Together with #341, this PR supersedes #332.

# Description

The announcement banner text html was added in #341. This PR now points
our
config to it, allowing the banner to be displayed.

Note: #332 also has a CSS fix. However, that fix is included in the
latest
release of the napari-sphinx-theme, so I think it's best to leave it out
of
this PR. (I would be happy to make a separate PR for it if we don't want
to
wait until the dependencies are updated.)

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>