Add NAP for multicanvas (#249)
Particular thanks to @brisvag and @melonora for previous work and
discussions on this, and for agreeing to be co-authors. I am happy to
absorb any criticism and direct praise their way 😄.

Thanks also to @andy-sweet and @Czaki, both of whom provided helpful
discussion while I worked on this NAP.

Of course I expect many changes to this and look forward to feedback.

---------

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>
Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>