Remove scaling of highlight width, fix showing shapes for uneven coordinates (#6629)
# References and relevant issues

This PR is created as result of debug
https://github.com/napari/napari/issues/6627#issuecomment-1914804363
As 0.4.19 branch has fewer problems than #6628 is created for 0.4.19. 

# Description

Fix logic for selecting of visible shapes based on idea from #2289. It
is lucky that we do not hit this earlier.

Removes scaling of width of markers border (this is same like in #6628)

To improve readability, this update `ShapesVisual` to have named
subvisuals.