Discard duplicate edges and interior triangles
This is a partial fix for #5673. Current limitations:

- only works when `triangle` is installed (`py-triangle` in conda-forge)
- currently on macOS, this works fine running in IPython but it segfaults when called from plain Python. I don’t know whether it’s just an env thing.

Summary:

It turns out that what we actually need is a *constrained Delaunay triangulation*[1], that is, a Delaunay triangulation that preserves specific edges from the vertices. That’s what the 'p' option in `triangle.triangulate` does *but* it doesn’t work if the edges are actually not planar — that’s why the triangulation was failing for this polygon even if triangle was installed. This PR fixes that by finding edges that appear twice in the polygon and removing them.

Additionally, certain triangles from the resulting triangulation need to be removed, and this is easy to do by running `skimage.measure.points_in_poly` on the triangle centers, which are fast to compute with NumPy indexing.

[1]: http://en.wikipedia.org/wiki/Constrained_Delaunay_triangulation