Ensure consistent layer tooltip (#6589)
# References and relevant issues
Closes https://github.com/napari/napari/issues/6435

# Description
In this PR I ensure that when mousing over a layer in the layer list the
tooltips are consistent.
The layer name is always shown, which covers the case of long names
which might be elided in the list (the linked issue). If there is a file
name and the other info, it's also shown. Here's an example of a long
layer name (but all layer names are shown):
<img width="357" alt="image"
src="https://github.com/napari/napari/assets/76622105/21829185-1981-4dab-9147-acbd9efe8b99">

Also, I add a small amount of logic to format the tooltip better, so
that when something is missing the spurious punctuation seen on live,
e.g. with empty Points, is eliminated.
Live
<img width="266" alt="image"
src="https://github.com/napari/napari/assets/76622105/7ac8a680-ad14-4106-83ae-0d221fe51f98">
this PR:
<img width="275" alt="image"
src="https://github.com/napari/napari/assets/76622105/290457ae-408e-43cd-ac91-b910efbd1b55">

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>