Fix scheduled benchmarks to test both last release and main (#6496)
The scheduled benchmarks are broken. In #6278 I fixed running benchmarks
against the main branch. But I missed that benchmarks must also run
against the last release (currently 0.4.19).

This PR adds checking benchmarks against the latest release in
pull_request tests.

For any API that was changed, this PR adds a conditional `if` branch and
uses the old api for 0.4.19 and the new one for main. This should
simplify after removing obsolete code in 0.5.0 release.

This PR enables Async benchmarks, but only on the main branch. 

I also renamed the previously misspelled `Skiper` class to `Skip` and
added constructor arguments to provide predicate functions about *when*
to skip:

* `if_in_pr`: used to skip benchmarks if the `PR` environment variable
is defined.
* `if_in_ci`: used to skip benchmarks if the `CI` environment variable
is defined. (Because CI has reduced RAM compared to most machines.)
* `always`: used to skip incomplete benchmarks, or to skip last-release
benchmarks on new functions.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>