Fix some weird behaviors when using delete and backspace keys as shortcuts (with or without modifiers) (#6406)
# References and relevant issues

Closes #6405 

# Description

* Allows using Delete and Backspace keys with modifiers as shortcuts
(for example `Ctrl + Delete` or `Ctrl + Backspace`) with different
OS/bindings combinations:


![delete_backspace_shortcuts_mod](https://github.com/napari/napari/assets/16781833/896652a4-ccda-43c4-8f3d-ef5f7f81df4f)

* Adds/improves tests for the `ShortcutEditor` widget.

## Notes:

* For a new test, I ended up using a new dependency (`pyautogui`) since
doing the test with the usual `QTest`/`qtbot` functions made it pass
while still showing weird behaviors when checking manually on
macOS+PyQt6. Probably, that is due to `QTest`/`qtbot` functions work
generating instances of `QKeyEvent` directly while with `pyautogui` the
actual OS API is used to generated the keyboard events. However, the new
test can't run on macOS CI (due to accessibility permissions which
causes it to timeout). ~Currently looking for an alternative for this:
maybe including the new test could not be worthy at all, skipping it on
macOS CI, or maybe I will end up finding a way to check things with the
usual `QTest`/`qtbot` functions.~ End up just creating a mark to skip
test on macOS CI and using it for the test

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>