Add viewer keybinds to select & show only next/prev layer and to toggle visibility of unselected (#6590)
# References and relevant issues

Closes: https://github.com/napari/napari/issues/2166
Closes: https://github.com/napari/napari/issues/5253

# Description

As requested, this PR adds keybindings to enable selecting and showing
only the next/previous layer in the layer list. By default these are
bound to Shift-Option-Up and -Down (see note below)
Additionally, to complement this and the existing `Toggle visibility of
selected layers` (default `V`), I added a keybinding for `Toggle
visibility of unselected layers` (default shift-V, because alt-V clashed
on Windows).

Edit: to prevent skipping layers due to Up/Down firing double, I capture
Up/Down with modifiers from QtLayerList so they are handed by the viewer
only and not by Qt. The alternative is to not do that and not have the
keybinding advance the selection, but this only works if Up/Down are
used, but someone might want to use W/S or some other non-arrow binding,
which would make it fail.

Note: it's more logical and natural feeling to have the `Select & show
only next/prev` be just Alt-Up and Alt-Down instead of Shift-Option-Up.
However, Alt-Up and Alt-Down change the focus of sliders. I suspect this
is a much less frequently used set of keybinds, so would be tempted to
change it to use Left Right, since left right keys are already are used
for manipulating sliders, it makes it more intuitive for up/down to
manipulate layer list and left/right to manipulate sliders.
Shift-Up/Down can't be used either, because it's multi selection.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>