Update docstring for _get_console method of QtViewer class (#6723)
# Description
The docstring for the `_get_console` method of the `QtViewer` class is
unfinished. I've made some minor changes to tidy it up.

## Context
I had a conversation with Juan about documenting return values. Most of
the other methods in the QtViewer class are also pretty sparse on
information (so another PR would be welcome), but this is the only
obviously unfinished one (so that's why this PR addresses it
specifically).

---------

Co-authored-by: Lucy Liu <jliu176@gmail.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>