Fix sync slicing of not yet visible layers (#6766)
# References and relevant issues
Closes #6651, #6760 

# Description
This changes how we handle sync slicing of non-visible layers in the
`LayerSlicer` by always slicing all layers using `Layer._slice_dims`.
The effect is that `Layer._slice_input` should now always be up-to-date
with the state of `ViewerModel.dims`, so that if the layer becomes
visible later (without the `Dims` state changing) it has the state
necessary to fetch the right data. If the layer is not visible, the
logic of `Layer.refresh` ensures we don't actually fetch data since that
would be wasteful.

Note that the logic of async slicing's `refresh` is fundamentally
different to sync slicing because all it does is notify the
`ViewerModel` that it should be refreshed because it assumes it does not
have all the state necessary to do slicing (because that's in `Dims`).
While that has some disadvantages, I still prefer that logic over the
current sync slicing approach, which is to keep the slice state from
`Dims` up to date with the corresponding state in the layer, so that
it's ready to fetch the right data in the future.