Lock dimensions / axes while rolling (#5986)
# Description


The ability to roll through dimensions is a useful functionality in
`napari`. However, in many use cases, it is desirable to only roll
through or reorder the spatial dimensions. Particularly when dealing
with time-resolved data, the time dimension should remain fixed most of
the time. Currently, achieving this requires manually reordering the
dimensions, which can be a tedious process.

## Original Behaviour
In the current version, all dimensions are in the reorder dimensions
pop-up window are draggable, and users can freely reorder or roll
through them.

## New Behaviour
The dimensions in the reorder dimensions pop-up window have a lock
symbol in front of there name. If clicked the lock closes and the item
is no longer draggable. When rolling these dimensions will stay in place
while the other dimensions are rolled.

[lock_dimensions_demo](https://github.com/napari/napari/assets/56394171/8caeff74-b437-45bf-9ee2-c2967c0372f5)

# References

- vscode
- https://doc.qt.io/qt.html#qtforpython
- https://github.com/napari/docs

## Type of change

- [x] New feature (non-breaking change which adds functionality)
- [x] Enhancement (non-breaking improvements of an existing feature)
- [x] Documentation (update of docstrings and deprecation information)
- [x] This change requires a documentation update

# How has this been tested?

- added tests for all newly introduced / modified functions and classes
- added test for new functionality of `QtDimsSorter`
- reworked old tests of `QtDimsSorter` to match new structure 

## Final checklist:
- [x] My PR is the minimum possible work for the desired functionality
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my
feature works
- [x] If I included new strings, I have used `trans.` to make them
localizable.
For more information see our [translations
guide](https://napari.org/developers/translations.html).

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Peter Sobolewski <76622105+psobolewskiPhD@users.noreply.github.com>
Co-authored-by: Peter Sobolewski <pete.sd@gmail.com>