Implement validation of docstrings and kwargs order (#6782)
# References and relevant issues
requested
https://github.com/napari/napari/pull/6779#pullrequestreview-1963287570

# Description

This PR adds two test utils functions. 

First is to check if all functions are mentioned in docstring (and if
there is not abandoned attribute in docstring). Then it checks if all
parameters are in the same order in both signature and docstring.

The second function checks if keyword-only arguments are declared in
alphabetical order.

Both are used for Image layers as it is sorted in #6428. It could be
applied to #6779 and other future PR.

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>