Improve repo configuration based on scientific python suggestions (#6554)
# Description

The Scientific python has tool named repo-review with set of rules that
they think are good for projects.

Not all apply to napari, but some of them looks usefull. To see the
check on napari repo one may use this link

https://learn.scientific-python.org/development/guides/repo-review/?repo=napari%2Fnapari&branch=main

This PR contains non mypy improvements, as mypy ones requires
improvements in code.

It depends on #6557

This PR also enables automatically dump graph when QtViewer is leaked
during tests. Then such graph in pdf is uploaded as job artifact.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Genevieve Buckley <30920819+GenevieveBuckley@users.noreply.github.com>