Bump python from 3.9 to 3.12 in pre-release tests (#6826)
# References and relevant issues

Spotted in https://github.com/scikit-image/scikit-image/issues/7385: as
many packages follow
[SPEC0](https://scientific-python.org/specs/spec-0000/), they have
already dropped 3.9 and may soon drop 3.10 from their builds. This means
that we cannot detect pre-release issues if those packages update and we
are only testing on 3.9. This PR bumps the pre-release testing to the
latest Python version as of this writing.

# Description

Use python 3.12 in pre-release tests. Drop pyside2 as it does not
support never python.