Fix selection size when canvas limits are set for Points (#6853)
# References and relevant issues
Closes https://github.com/napari/napari/issues/6682

# Description
The selection markers were not updated accordingly.

This is still not perfect (zooming in and out beyond the limits will
change the thickness of the highlight), but it's a limitation of how
hilight is implemented (by just using extra markers). Ideally we want to
have a single marker take care of hilight as well, but this requires
rewriting the shaders and it's not trivial.