Fix citation cff and add worflow for future validation (#6888)
# Description

We have a problem that there is no entry on zenodo for napari for 0.4.19
release line. It lokks like some PR (most probably mine) broke the
citation cff format.

I have validated and fixed field based on `cffconvert` python package. I
also added GitHub actions for validate file content based on official
documentaion
https://github.com/citation-file-format/citation-file-format/blob/main/README.md

I'm not 100% sure that it will be enough to fix it, but it looks
required for that.