refresh when canvas_size_limits are changed (#6834)
# References and relevant issues
If you change `canvas_size_limits` programmatically, nothing happens.
You need to interact with the canvas. 
This is a bit confusing when using a notebook, ipython, etc. because it
appears it doesn't work or the size limits are not correct.

# Description
Uses `self.node.update()` in ` _on_canvas_size_limits_change`

---------

Co-authored-by: Lorenzo Gaifas <brisvag@gmail.com>