"Shapes to Labels", "(float)Image to Labels": Don't pop layer being converted, append new one (#6859)
# References and relevant issues
When you use the layerlist contextual menu to convert a layer, e.g. to
labels, it pops the layer.
I find this to rarely be the behavior I want. Worse yet, **there is no
easy undo of that** the layer is just gone! I feel like being
conservative is better -- I can always remove the previous layer myself
if I truly con't want it.

# Description
This PR drops the `pop` of the Shapes layer from the LayerList and
instead the new Labels layer is appended above the layer being
converted. I updated the test accordingly.

The action calls a private function which isn't used anywhere else, so I
don't think it needs any deprecation warnings. The existing docs are
vague enough that there is no need to update anything.
But I can do so, if folks think it's useful.

Additionally, because images with float data are also converted to int
before convert to labels, I made that also not `pop` the original image
layer, because it is not fully reversible.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>
Co-authored-by: Juan Nunez-Iglesias <jni@fastmail.com>