Fix use of scaling in vispy points (#6894)
# References and relevant issues
In #6025, the vispy update broke the efforts from #5312. This is
actually because the changes in vispy apparently do not handle the
`scaling` kwarg properly at initialization. We can fix it here by not
passing it.

I can't believe this has been wrong for so long, and I somehow never
connected this issue that kept bugging me with that PR... but here we
are.

# Description
Try the following on main and here and move the camera around:
```py
import napari
import numpy as np
v = napari.Viewer(ndisplay=3)
pl = v.add_points(np.array([[1, 1, 1], [10, 10, 10]]), scale=[1, 5, 5])
pv = v.window._qt_viewer.canvas.layer_to_visual[pl]
```
This fixes two things:
- point sizes being all messed up by scaling
- the anisotropic scale rotation wobblyness you see above

This returns to the correct, pre-#6025 situation.