Automatically remove run-benchmarks label once benchmarks are executed (#6929)
# Description

Iterating on a performance PR that needs benchmarks is currently more
cumbersome than it needs to be: add the run-benchmarks label, get a
performance report, act on the report (ie make code changes and push),
then, to run the benchmarks again, a core dev needs to *manually remove*
the label, then immediately add it again. Instead, this PR automates the
removal of the label once the benchmarks are done running and the report
has been generated. In fact, it should also remove the label even if the
benchmarks fail due to some bug in the code or benchmark code.