Fix array data copy behavior for numpy 1 and 2 consistency (#6932)
# References and relevant issues

Failure in our prerelease tests which picked up `numpy==2.0.0rc2`:
https://github.com/napari/napari/issues/6871#issue-2263425683

Related discussion on Zulip:
https://napari.zulipchat.com/#narrow/stream/212875-general/topic/handling.20the.20numpy.202.2E0.20release

# Description
This fixes the coercion of data using the `Array` pydantic custom class
so that it works for both numpy 1 and 2, and only makes a copy if
needed.

Instead we could consider using `np.asarray`, but we'll also need to
replace the `ndmin` argument (which `np.asarray` does not have), by
using `np.expand_dims` or similar. I have done that it in one other
place where I don't need those fancy arguments.