Update CITATION.cff with contributors for 0.5.0 (#6958)
Thanks to @Czaki for reminding me that there's no need to do this
manually when he created a
[script](https://github.com/napari/napari-release-tools/blob/main/find_contributors_without_citation.py)
to get this info. 😅

**Edit:** categorising into not responded/added/declined:

Not yet responded:
@AmirAflak @DanGonite57 @mstabrin 

Added:
@AndrewAnnex @MarchisLost @aeisenbarth @imagejan @jamesyan-git
@jo-mueller @maweigert @odinsbane @palec87 @perlman

Declined:
@stefanv

Thank you for contributing to napari! If you would like to be included
in the CITATION.cff file for the next release, please comment by filling
in the following template (in a code block) and I'll add your response
to the file:

```
- given-names:
  family-names:
  affiliation:
  orcid: (optional, format as url: https://orcid.org/0000-0000-0000-0000)
  alias: (gh handle)
```

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>