Block zarr 3.0.0a0 on pre tests (#7028)
# References and relevant issues
closes https://github.com/napari/napari/issues/6997
replaces #6999 (apologies Czaki I cannot push to your branch or PR to
your repo - feel free to port these changes to your branch if it works
for you)

# Description
This PR blocks using `zarr==3.0.0rc1` in pre tests by add it to
`resources/constraints/version_denylist.txt` and use this file as
constraints for pre tests.

This also fixes a failure in the --pre macOS-latest/pyqt6 workflow. My
suspicion here is that it is not properly waiting for some of the macOS
fullscreen animations, and it causes the app/testing environment to lose
focus. Unfortunately I could not find a better way to fix it than just
waiting. Any signal/condition I waited for did not seem to work. I guess
they happen before the animation runs.

Initially I narrowed the cause down to this test by running a [modified
version of
`detect-test-pollution`](https://github.com/aganders3/detect-test-pollution)
in CI [on my fork](https://github.com/aganders3/napari/pull/17), and
eventually I was able to replicate this locally. I'm still not sure why
it's only a problem in macOS-latest/pyqt6.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>