[Enh] Labels polygon tool: use a setting for completion_radius, by default double-click always works. (#7023)
# References and relevant issues
Another part of addressing https://github.com/napari/napari/issues/6489

# Description
In this PR I add a setting under Settings > Experimental for the labels
polygon tool `completion_radius`. This is the maximum radius from the
first vertex for double-click completion of the shape to work.
By default, I set this now to `-1` by default, which makes double-click
always complete.
I clarified the bool variable controlling the behavior, renaming to
`use_double_click_completion_radius `.
Then, in the overlay machinery, I add a method to get the value from the
settings and if it's negative, disable using the completion_radius, so
double-click to complete *always* works, just like for the Shapes layer.

For a given Labels layer, you can still/now do:

`viewer.layers['Labels']._overlays['polygon'].use_double_click_completion_radius
= True`
and then it will return the old behavior with the default radius of 20.
You can do:
`viewer.layers[0]._overlays['polygon'].completion_radius = 50`
to change it.

In the process I discovered that the last vertex was being deleting. e.g
it wasn't possible to make a triangle.
This was masked by the fact that last vertex had to be so close to first
vertex, but now that you can double-click to complete anywhere it
becomes apparent. So I've fixed that too, keeping the vertex from the
double-click location, just like the Shapes layer polygon tool.