[Shapes] for line tool, have _fixed_aspect (shift) enable 45 degree rotations if shift held first. (#7043)
# References and relevant issues
Closes https://github.com/napari/napari/issues/7042

# Description
In this PR I first ensure that when drawing a new shape the aspect ratio
is reset to 1.
Previously, because a shape could be selected when drawing is begun
(e.g. drawing multiple lines back to back), the wrong aspect ratio (that
of the previous shape) would be returned by layer._aspect_ratio when
Shift is pressed. This had no ill effects, because it wasn't used with
shape drawing.
Next, with this information, I can check whether a line was being drawn
with Shift already pressed (_aspect_ratio ==1 ) and if so, allow it to
rotate through fixed 45 degree angles. If Shift is pressed *during*
drawing, the current behavior is maintained, where the current aspect
ratio is fixed.
I added a test for the new behavior.