Provide default colors automatically in DirectLabelsColormap
While trying to fix ome/napari-ome-zarr#109 on the ome side after #7025,
I found that I was still getting errors — now by the color_dict
validator, which requires a None key or a defaultdict. Ultimately,
I think this is a quirky interface that I don't want to impose on
ome-zarr or their readers. Therefore, I've changed the ValueError to a
warning. I'd consider removing the warning altogether but I think
@Czaki was opposed?