Fix vispy pixel offset in 3D multi-scale image layers (#6319)
# References and relevant issues
See #6320 

# Description
This solves a specific problem in #6320 around 3D display of multi-scale
image layers consistently with other data (e.g. 3D points data). As
napari currently only ever shows the lowest resolution of a 3D image, we
use the corresponding multi-scale offset to ensure that vispy starts
drawing the 3D object at the intended location so that display is
consistent between 2D and 3D, and with other layer types (e.g. `Points`
is an easy one to test).

---------

Co-authored-by: Andy Sweet <andrew.d.sweet@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>