Restore events to `QtViewer.canvas` (#7060)
# References and relevant issues

In #7054, I noted that in #5432 we had inadvertently removed access to
mouse events from QtViewer.canvas, without a deprecation. @jni
[commented](https://github.com/napari/napari/issues/7054#issuecomment-2205548968)
that we could temporarily restore access with a property, followed in
later versions by an alternate API and deprecation.

# Description

This PR adds an `events` property to QtViewer.canvas, restoring the
broken behavior. It does not yet have a deprecation message because we
don't have an alternate API.