Ensure out-of-bounds layers are loaded (#7077)
Closes #7070

In #7070, @psobolewskiPhD reported that layers that are out of bounds of
the dims slicer never appear loaded. This caused the layer thumbnail
spinner to spin forever, which is jarring to users — it makes it
impossible to know whether there is no data on that slice or whether
some other error (e.g. slow network) is causing the layer to not load.

Some investigation showed that the problem comes from out-of-bounds
layers using _ImageSliceResponse.make_empty, which always returns a new
slice ID — never the slice ID that was requested. When the layer checks
whether the returned (correctly empty) slice is the slice it requested,
it finds that it isn't, so it does not change its state to loaded.

Always returning the requested slice ID was not a good option because
empty slices are also used on layer construction, before any data is
requested. In those cases, returning a new slice ID is the correct
behaviour.

This PR adds a new, optional "request_id" argument to
_ImageSliceRequest.make_empty, and sets it correctly when requesting an
out of bounds slice for an image.

I've added a new test for this behaviour and have confirmed that it
fails on main:

```
>           qtbot.waitUntil(partial(layer_loaded, l1), timeout=50)
E           pytestqt.exceptions.TimeoutError: waitUntil timed out in 50 milliseconds

test_async_slicing.py:70: TimeoutError
```

But passes with the changes in this PR.