Fix logic to keep settings and preference dialog widgets in sync (`dask`, `highlight`, `brush_size_on_mouse_move_modifiers`, `new_labels_dtype` settings errors) (#7027)
# References and relevant issues

Closes #7020 

# Description

Remove logic that overrides dask settings event handling and update
`QtViewer` logic to handle the change and handle nested event model
settings changes to update preferences dialog widgets.

As a little bit of context, due to the changes done at #6113 now there
is some logic to keep synced the settings and the preference dialog
generated widgets. This logic relies on the settings events. With the
way dask settings was working, triggering the `enabled` checkbox for
dask cache over the preference dialog was in turn emitting an event from
the applications settings emitting a change for the dask setting object
itself which then was being passed as a value to be set over the
preference dialog widgets which is something they can't handle (the
preference widget for objects expects as state a dictionary as far as I
understand)

# Note:

* While checking this I also noticed issues with `Brush size on mouse
move modifiers` and `New labels data type` settings due to the usage of
`StrEnum` for those settings values
* With the change to the sync logic to also react to nested settings
done here some changes needed to be done to take into account
individually changing highlight settings (for example changing the color
but not the thickness)
* Checking the `Restore defaults` logic seems like after resetting the
preference dialog widgets was being rebuild. Thinking about it seems
like that is not necessary anymore since the widgets can react to
changes over the settings now. For the moment just commented the line to
check how thigns work without rebuilding locally but maybe something to
check on a different PR?
* Still double checking if the fix and test added here are enough but
any feedback is appreciated!