Bump the settings schema version (#7053)
Closes #5316

In #5103 the keybinding model was moved from an internal implementation
to app-model's implementation. This changed the representation of
keybindings on disk, *and* it changed it in a platform-dependent way: on
Mac, shortcuts that use the Command key are now written as Meta-Key,
where they used to be translated to Ctrl-Key so that the preferences
file could be platform-independent.

This is fixed with a migrator that swaps the control and command keys
when reading keybindings from a file using the 0.5.0 schema
(confusingly, the schema used in napari 0.4.19) but leaves them
untouched if they come from defaults or from a 0.6.0 schema version
file.

Additionally, app-model's keybinding validator does not recognise some
previously-valid keys, including Control. This was *supposed* to be
address by a string substitution in our shortcuts validator, but it was
run *after* the keys had been validated by app-model's keybinding class,
which had cleared them as invalid. This PR fixes that issue by adding
`pre=True` to the pydantic validator declaration, which makes the
validator run with the raw input rather than the leaf-validated input.

Finally, I added a couple of key translations to make it easier for
people to hand-write keybindings: Super, Command, and Cmd are all valid
modifier descriptions now.