Use `viewer.layers` instead of  `_layers.model().sourceModel()._root` for dummy context creation (#7109)
# References and relevant issues

# Description

Prior to this PR we were retrieving the `LayerList` using a private
attribute, but there is an equivalent public attribute, and it's unclear
what the original motivation was behind using the private version. This
PR simply updates the two locations where
`_qt_viewer._layers.model().sourceModel()._root` was used to instead use
`_qt_viewer.viewer.layers`. This should be less fragile.