Always add `Empty` context key, even if `action` is already registered (#7106)
# References and relevant issues

Closes #7105 

# Description

This PR ensures that even if the dummy `Empty` action is registered with
the `app`, we still assign its `context_key`. I checked that there's no
danger in assigning the context key multiple times (like there is in
registering an action multiple times), so this shouldn't have any
undesirable side effects... 😬

Because new viewers in notebooks share the same `app` instance, the
`Empty` dummy action is already registered with all launched viewers,
but each launched viewer gets its own instance of the `LayerList` and
therefore its own new instance of the `context`. This means we **keep**
the registered actions, but have to reassign all the context keys. I
think this is terrible, but should be fixed in concert with whatever we
decide in #7101.

---------

Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>