[Bugfix] Fix logic in setting settings using env vars. (#6669)
# References and relevant issues
Closes: #6637 

# Description
This PR fixes the logic in the `nested_env_settings` method to:
1. ensure that for each sub-field the env vars are set, but nesting all
of the subfield related checks and setting within the subfield for loop
2. and also makes sure to not break loop too early, such that more than
one env var could be set.

I added tests that fail on `main`, plus a sub-cases that passes on main
but wasn't tested.

Edit: berg, the diff is annoying. 
The first change is just indenting all the logic under `for subf in
field_type.__fields__.values():`
The 2nd change is removing the early break in the `else` of `for e in
subf.field_info.extra.get('env_names', []):`. This break in the for loop
else was breaking the parent loop early.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Grzegorz Bokota <bokota+github@gmail.com>