fix condition of adding asterix, when save reader of path (#7112)
# References and relevant issues

# Description
Current logic for saving a preference from the reader dialog leads to
adding `*` to absolute paths, making it impossible to save a reader
preference for a zarr directory. This was because an asterisk was added
whenever the opened path had an extension.

This PR fixes checking if the path is an absolute path and doesn't add
an asterisk if it's a directory.